Changed package names to describe the optimization more clearly
