+++ src/main/java/me/jellysquid/mods/lithium/mixin/entity/collision_only_boat_shulker/MixinWorldChunk.java
@@ -1,4 +1,4 @@
-package me.jellysquid.mods.lithium.mixin.entity.use_classed_collision_box_retrieval;
+package me.jellysquid.mods.lithium.mixin.entity.collision_only_boat_shulker;
 
 import me.jellysquid.mods.lithium.common.world.ExtendedChunk;
 import net.minecraft.entity.Entity;
