+++ src/main/resources/lithium.mixins.json
@@ -42,23 +42,23 @@
         "chunk.serialization.MixinPalettedContainer",
         "collections.entity_filtering.MixinTypeFilterableList",
         "entity.block_cache.MixinLivingEntity",
+        "entity.collision_only_boat_shulker.MixinBoatEntity",
+        "entity.collision_only_boat_shulker.MixinLithiumEntityCollisions",
+        "entity.collision_only_boat_shulker.MixinShulkerEntity",
+        "entity.collision_only_boat_shulker.MixinTypeFilterableList",
+        "entity.collision_only_boat_shulker.MixinWorld",
+        "entity.collision_only_boat_shulker.MixinWorldChunk",
         "entity.collisions.MixinCollisionView",
         "entity.collisions.MixinEntity",
         "entity.collisions.MixinEntityView",
         "entity.data_tracker.no_locks.MixinDataTracker",
         "entity.data_tracker.use_arrays.MixinDataTracker",
-        "entity.shortcut_entity_entity_collisions.MixinEntity",
         "entity.fluid_checks.MixinEntity",
-        "entity.use_classed_collision_box_retrieval.MixinBoatEntity",
-        "entity.use_classed_collision_box_retrieval.MixinLithiumEntityCollisions",
-        "entity.use_classed_collision_box_retrieval.MixinShulkerEntity",
-        "entity.use_classed_collision_box_retrieval.MixinTypeFilterableList",
-        "entity.use_classed_collision_box_retrieval.MixinWorld",
-        "entity.use_classed_collision_box_retrieval.MixinWorldChunk",
-        "entity.type_specific_entity_retrieval.MixinAbstractDecorationEntity",
-        "entity.type_specific_entity_retrieval.MixinArmorStandEntity",
-        "entity.type_specific_entity_retrieval.MixinFormCaravanGoal",
-        "entity.type_specific_entity_retrieval.MixinItemFrameEntity",
+        "entity.replace_entitytype_predicates.MixinAbstractDecorationEntity",
+        "entity.replace_entitytype_predicates.MixinArmorStandEntity",
+        "entity.replace_entitytype_predicates.MixinFormCaravanGoal",
+        "entity.replace_entitytype_predicates.MixinItemFrameEntity",
+        "entity.stream_entity_collisions_lazily.MixinEntity",
         "math.fast_util.MixinAxisCycleDirection",
         "math.fast_util.MixinBox",
         "math.fast_util.MixinDirection",
